Tab-Switching funitonality working (Almost!)
The tabs switch and display tasks correctly. Small bug with completion status being displayed incorrectly. I chose to scrap the idea of sorting by different options in one function that resided in the List class because he function returned an array and not a ToDo object which gave me issues. I decided to break it up into smaller boolean functions that  determined if an element in the list would be displayed or not as a for loop went through the list. This was much easier and cleaner. Will fix bug and begin working on the functionality to delete a task.